+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changes to the SDK
-
+### Version 0.1.2-beta
+- Fix objectid bug in Core.getfileliststatus and Core.getaclstatus and add unit test for it
+- Remove blocksize and replicationfactor from directory entry
+- Change parameter names for some public APIs
 ### Version 0.1.1-beta
 - Reference Microsoft.Rest.ClientRuntime instead of Microsoft.Rest.ClientRuntime.Azure.Authentication in Microsoft.Azure.DataLake.Store project.
 - Reference Microsoft.Rest.ClientRuntime.Azure.Authentication in Microsoft.Azure.DataLake.Store.UnitTest.